+++ src/main.js
@@ -14,5 +14,3 @@ document.querySelector("footer").innerHTML = footer.template;
 enrutador.observadorRutas();
 // Cargamos la página home
 window.location = "#/home";
-
-ls.setUsuario({ email: 'chafardera@gmial.com', rol: 'registrado' })
